Ensure the DEP enrollment profiles get regenerated when enable release changes
