+++ server/mdm/apple/apple_mdm.go
@@ -91,17 +91,16 @@ type DEPService struct {
 // getDefaultProfile returns a godep.Profile with default values set.
 func (d *DEPService) getDefaultProfile() *godep.Profile {
 	return &godep.Profile{
-		ProfileName:           "FleetDM default enrollment profile",
-		AllowPairing:          true,
-		AutoAdvanceSetup:      false,
-		AwaitDeviceConfigured: false,
-		IsSupervised:          false,
-		IsMultiUser:           false,
-		IsMandatory:           false,
-		IsMDMRemovable:        true,
-		Language:              "en",
-		OrgMagic:              "1",
-		Region:                "US",
+		ProfileName:      "FleetDM default enrollment profile",
+		AllowPairing:     true,
+		AutoAdvanceSetup: false,
+		IsSupervised:     false,
+		IsMultiUser:      false,
+		IsMandatory:      false,
+		IsMDMRemovable:   true,
+		Language:         "en",
+		OrgMagic:         "1",
+		Region:           "US",
 		SkipSetupItems: []string{
 			"Accessibility",
 			"Appearance",
@@ -207,6 +206,11 @@ func (d *DEPService) RegisterProfileWithAppleDEPServer(ctx context.Context, team
 	// ensure `url` is the same as `configuration_web_url`, to not leak the URL
 	// to get a token without SSO enabled
 	jsonProf.URL = jsonProf.ConfigurationWebURL
+	// set await_device_configured based on appCfg/team enable_release_device_manually
+	jsonProf.AwaitDeviceConfigured = !appCfg.MDM.MacOSSetup.EnableReleaseDeviceManually.Value
+	if team != nil {
+		jsonProf.AwaitDeviceConfigured = !team.Config.MDM.MacOSSetup.EnableReleaseDeviceManually.Value
+	}
 
 	depClient := NewDEPClient(d.depStorage, d.ds, d.logger)
 	res, err := depClient.DefineProfile(ctx, DEPName, &jsonProf)
